same as last update, just fixed screen-resize
i forgot to test screen-resize last time, so THIS is the actual last update to "pygamesim old" (which had the same bug). I changed the screen size variable ("drawSize") to a tuple several updates ago, but never tested the video-resize, which had code that tried to edit entries of that tuple, which is of course illegal.