fixed zooming bug
there was a flag (boolean) that indicated whether the car polygon/sprite has to be recalculated, to save a little processing power, but i removed it because it only works for the car-centered camera.
this update just removes a single leftover from that system.
soon to come:
new version of the program, now simply called "coneConnecting.py", which separates the functional class and the drawing class. This is to allow the functional class to be used separately. After that, i'll make a version (or just a class within the main sketch) that lets the variables found in the functional class be streamed in (over IP network?), sothat the functional program can be on an entirely different PC. That will require some coordination with the transmitting side though, so it might take a while for that to be ready.