fixed important path-finding bug
path finding bug: it check whether the cone it intends to pair up (with an opposite one) is already connected with a (samecolored) cone that has already been used to make a path, because you don't want the path to start progressing backwards, but it was checking only left cones for this. Now it checks both sides correctly
this will be the last update to pygamesim, as i'm now moving to coneConnecting.py, which has the functinal and drawing parts seperated

to be clear:
coneConnecting.py is the newest (and last) namechange in this folder. it contains 4 classes (so far): coneConnecter (which holds all the functional math for connecting cones, and can be used in other programs), pygameDrawer (which holds all the drawing functions, just for visualization), pygamesimLocal (which combines the last two classes as subclasses, this is the replacement for the pygamesim class) and lastly, raceCar (which is a (temporary) class for the car, and holds the functions that make it drive and such. The drawing functions were moved from raceCar() to pygameDrawer, but still use the length & width data from raceCar() objects)

also, little quality of life improvement: holding shift while placing cones will make them attempt to connect upon creation (as opposed to later/when clicked on again)
i stopped connecting cones upon creation out of fear it would make the job too easy, as all other cones on the screen are already connected, it only has 1 option, and it's pretty much always going to be placed by a human, who would place it in order of how the track progresses (so too easy). But i found that if i'm in a hurry and just want to plop down some shit, it's nice to be able to hold shift and get it done