(start of?) clock rework
by having the clock function as global (outside the map class), and just setting the __default argument to the starting time, it meant that the starting time was reset every time a new map object is created (anywhere), which can happen a lot, with mapSending and file loading. clockSet can therefore no longer be used by default in __init__, and will be removed in a later version.