fixed clock issues
pickle wanted to copy the entire pygamesimLocal object, becuase it houses the clock function (which is passed to the car object). I fixed this by letting pickle copy only the internalClock function in the new (fresh) map object created in shallowCopyExtractMap(). this does mean that special clocks may not work as intended, when viewing through a remote window, but that is only for simulations anyway, so it shouldnt matter too much.