added map printing functions
every class in Map now has a __repr__() function, which is called whenever python tries to print() an object of that class. Basically, instead of printing "<Map.Cone object at 0x000001D26820D160>" or whatever, it now prints like "Cone(ID=1,LEFT,pos=(1.0,2.0),FINISH)". Same goes for Targets, Cars and the Map object itself.
the downside of doing this for the Map object itself, is that anything that inherits the Map class now also inherits the __repr__ function (unless the inheriting class overwrites it with a __repr__ of its own). That is why the Map.__repr__ also prints the type of 'self'.