first version of documentation
i though (and was instructed to by my Minor handler) it would be a good idea to have some more documentation. I still need to add sources (which is tricky, because i mostly just made stuff up) to make it more fancy.
each word doc corresponds to a python file.
all of this will also be on the Google Drive!