updated calibration and fixed bugs
i made a mistake in the serial input parsing, which made the timeout not work. Now, it actually does wait for data to come in (as it should).
i also changed the steering servo control, feedback and motorKpSteer to be polynomial instead of (the rather dodgy) if-statement linear/exponential functions. Currently all 3 curves are 2nd order polynomials, but i did experiment with 3rd order (and made some math-fixing-functions to avoid float overflow) if we wish to go there in the future.
i still don't know why the (mechanical) linkage between the steering servo and the wheels is nonlinear (or at least asymmetrical), and i'm still just assuming a wheelbase (as i have no accurate way of measuring).