major overhaul
(see updated README.md)
i made another new version because i'm stubborn and bored. It's once again singlecore and it runs buttery smooth on my laptop. I carved out most of the multicore slowdowns, but that also removed the remote map viewing (the pickling and network overhead is huge). I don't expect anyone to take note of this thing (it is and stays depricated), but i like it enough to replace the old (less good) stuff (which is publicly available and has my name on it...)